fix: return item requirements as int, not a byte
MItem::GetRequireSTR/DEX/INT/Level returned BYTE and GetRequireSUM
WORD, while the values they compute exceed a byte by design: the
table value plus twice the options' requirement sum is capped at
MAX_SLAYER_ATTR (295) for slayer gear and not capped at all for
Ousters gear. Anything from 256 up wrapped - a 295 requirement came
back as 39, a 300 one as 44 - so a level-150 item with strong
options looked easy to equip on the client, and the gear checks in
MCreature/MPlayer compared the player's stats against the wrapped
number. The functions return int now; every caller already stores
the result in an int or compares it with one.

Found while writing the item core's tests (task 4.4, second slice):
the first Ousters expectation came back as 4 for 260.

Test path: lib + test. RequirementsAboveTwoHundredFiftyFiveSurvive
fails before the fix (295 -> 39 three times, 300 -> 44) and passes
after; the in-range cases keep their own test. Suite: 216 tests
(3,513 checks) in the ordinary tree.

+++ Client/MItem.h
@@ -351,11 +351,13 @@ class MItem : public MObject, public CAnimationFrame {
 		//---------------------------------------------------
 		// 필요능력
 		//---------------------------------------------------
-		BYTE					GetRequireSTR()	const;
-		BYTE					GetRequireDEX()	const;
-		BYTE					GetRequireINT()	const;
-		BYTE					GetRequireLevel() const;		
-		WORD					GetRequireSUM() const;		
+		// int, not BYTE: the slayer ceiling is 295 and Ousters gear is uncapped,
+		// so a byte return wrapped level-150 requirements (task 4.4 fix).
+		int					GetRequireSTR()	const;
+		int					GetRequireDEX()	const;
+		int					GetRequireINT()	const;
+		int					GetRequireLevel() const;
+		int					GetRequireSUM() const;
 
 		//---------------------------------------------------
 		// 장착했을 때 바뀌는.. BasicActionInfo
